test(stateappevents): re-file TestColdStart* into coldstart test file
The 4 TestColdStart* funcs (testing coldStartFrom) lived in
snapshot_internal_test.go, breaking the 1:1 test<->source mapping. Move them to
coldstart_internal_test.go and relocate their shared helper writeContrivedSnapshot
(used only by coldstart + migrate) to helpers_test.go. Test-only.
